fix warning

+++ src/threadpool.cpp
@@ -23,7 +23,7 @@ namespace kmdiff {
   ThreadPool::ThreadPool(size_type threads)
   {
     if (threads < _n) _n = threads;
-    for (int i = 0; i < _n; i++)
+    for (size_t i = 0; i < _n; i++)
     {
       _pool.push_back(std::thread(&ThreadPool::worker, this, i));
     }
